+++ api/v1alpha1/doc.go
@@ -1,0 +1,2 @@
+// Package v1alpha1 defines the api boundary between CUE and Holos.
+package v1alpha1
