+++ docs/examples/cue.mod/gen/github.com/holos-run/holos/api/v1alpha1/result_go_gen.cue
@@ -1,0 +1,16 @@
+// Code generated by cue get go. DO NOT EDIT.
+
+//cue:generate cue get go github.com/holos-run/holos/api/v1alpha1
+
+package v1alpha1
+
+// Result is the build result for display or writing.  Holos components Render the Result as a data pipeline.
+#Result: {
+	#TypeMeta
+
+	#Kustomization
+	metadata?: #ObjectMeta @go(Metadata)
+
+	// Skip causes holos to take no action if true.
+	Skip: bool
+}
