+++ docs/examples/platforms/reference/clusters/projects/projects.cue
@@ -1,0 +1,3 @@
+package holos
+
+#InputKeys: project: "projects"
