--- docs/examples/platforms/reference/projects/secrets/components/namespaces/inputkeys.cue
@@ -1,8 +1,0 @@
-package holos
-
-#TargetNamespace: "external-secrets"
-
-#InputKeys: {
-	project: "secrets"
-	service: "eso"
-}
