--- docs/examples/platforms/reference/projects/secrets/components/namespaces/namespaces.cue
@@ -1,10 +1,0 @@
-package holos
-
-#InputKeys: component: "namespaces"
-
-metadata: name: #InstanceName
-objects: [
-	#Namespace & {
-		metadata: name: #TargetNamespace
-	},
-]
