--- docs/examples/platforms/reference/projects/secrets/components/namespaces/output.cue
@@ -1,6 +1,0 @@
-package holos
-
-// Output schema
-{} & #KubernetesObjects & {
-	ksObjects: [#Kustomization]
-}
