+++ docs/examples/project_types.cue
@@ -1,0 +1,11 @@
+package holos
+
+#Project: {
+	name: string
+}
+
+#Projects: {
+	[Name=_]: #Project & {
+		name: Name
+	}
+}
