+++ pkg/version/embedded/minor
@@ -1,1 +1,1 @@
-58
+59
