+++ pkg/version/embedded/patch
@@ -1,1 +1,1 @@
-3
+0
